Make the FileServer ops take pointer
Instead of structs.  Otherwise, we're passing the struct values, not the
object reference.  (Based on conversations with Russ).

There might be other problems like this, and also things we can clean
up.  For instance, the mutex doesn't need to be a pointer.  We were
getting away with the lack of a struct pointer since the mutex itself
was a pointer.  But things like the bool 'versioned' was busted.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>